requirements.txt
Django==3.2.5
asgiref==3.4.1
celery==5.1.2
django-celery-results==2.2.0
django-csvimport==2.16
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
install==1.3.4
psycopg2-binary==2.9.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
